Migrate deploy-commands to TypeScript
